Big rewrite of structure
Moved repeating code into new file and created a class that will convert
code between languages so that functions such as the If and While
calculators are far less clunky because they don't need to make a
different variable for each language with an exception. Started to tweak
encoding classes so that it is easier to write new ones. I plan to make
a template but I got GCSEs atm so not enough time.